Solve crash when using large RMS windows
Using relatively large RMS kernel sizes (e.g. a RMS kernel size (~50) on an image of size ~2k)
causes the error: "Kernel size > image dimension". This is because the Gaussian is evaluated
up to 40 sigma, and 40 x 50 x PSFsize > 2k, causing the convolution operation to throw an
error.

Fixed by truncating the Gaussian down to the image size if it extends beyond the image size.