Make image-based deconvolution work
With these changes it is possible to deconvolve an image, as long as no major iterations are
requested. Example parameters that can be added to test this are

wsclean -niter 1000000 -auto-threshold 3 -auto-mask 8 -multiscale ...

Important fixes are:
- When facetting, the PSF must be "processed" before it is stored in its CachedImageSet, because it will otherwise
  not be normalized.
- The deconvolution process now receives an imagingtable that has been 'collapsed' along the facet direction.
  Maik had already added a useful method for this that is now used (like `imagingTable.GetFacet(0)`).

On top of these changes, I came across some minor inconsistencies (e.g. still use of doubles for images) as well
as some small bugs in the deconvolution.

I haven't checked multi-frequency or multi-polarization cleaning yet.