Encode AVX-feature into package name
When building non-portable code (which is the default for performance
reasons), determine which AVX features are enabled by the compiler,
and encode the "highest" enabled feature into the package name.
This way, a user can determine from the package name if it makes sense
to install a package or not, depending on the target CPU architecture.