Fix setting of MPI compile flags/options
The current implementation uses 'add_compile_options()' to add
${MPI_COMPILE_FLAGS}. There are a number of issues with this:
  1) MPI_COMPILE_FLAGS has been long deprecated and is superseded by
     MPI_<lang>_COMPILE_FLAGS;
  2) MPI_COMPILE_FLAGS returns a white-space separated string of arguments,
     whereas 'add_compile_options()' expects a list (a semicolon separated
     string of arguments).
  3) MPI_<lang>_COMPILE_FLAGS have been deprecated as well, and CMake >= 3.10
     uses MPI_<lang>_COMPILE_OPTIONS instead. The latter returns a list, so
     it can be fed directly into 'add_compile_options()'
This patch provides a back-ward compatible solution. It checks if the variable
MPI_<lang>_COMPILE_OPTIONS is defined. If so, it uses 'add_compile_options()'
to add its value; otherwise it uses 'add_definitions()', which expects a
white-space separated string of arguments, to add the value of
MPI_<lang>_COMPILE_FLAGS.