Fix spectral fitting in the presence of missing frequency data
When a deconvolution channel contains no data, the residual image contains NaNs
and the weight of such a channel becomes zero. NaN value can never enter a
calculation, even not weighted, because even 0 x NaN is NaN. This is taken care
of in various positions, but wasn't yet properly dealt with in the spectral
fitter. This contains two fixes:
- Zero-weight values are excluded from the spectral fitting
- The used reference frequency is made sure to contain valid values, by
using the non-weighted frequency when all subchunks have zero weight.

This fixes #71 reported by Oleg Smirnov.