Temporary fix reservation for beam images
IDG computes the average beam upon making the PSF.
With the introduction of the direction dependent PSF,
the number of PSFs computed changes.

The mechanism for computing the average beams needs
to be adopted for the dd-psfs. For now just reserve
sufficient space for the average beams, so the application
at least does not crash when combining idg and dd-psf