PAN-95 Seperate gridding reads from processing
Split IO out of `GetCollapsedVisibilities()` and `GetInstrumentalVisibilities()` so that it is seperated from the data processing/logic.

A new method `ReadVisibilities()` is introduced that does the IO that was previously mixed in with the other operations.

Overloads of `GetCollapsedVisibilities()` that perform computations become `CollapseVisibilities()` Overloads of `GetInstrumentalVisibilities()` that perform computations become `ApplyWeightsAndCorrections()`

A single version of `GetCollapsedVisibilities()` and `GetInstrumentalVisibilities()` remain, but they are light inline wrapper functions that simply call `ReadVisibilities()` on the msReader and then one of `CollapseVisibilities()` or `ApplyWeightsAndCorrections()` and finally do the simple write that was in \`WriteImagingWeightsForRow(). Gridders have the potential option of calling all/some of these individual parts themselves instead if they need to do things differently (e.g. shared data reads).

By doing this we seperate IO operations from the logic leading to cleaner code with less side effects that will be easier to work on, write tests for and optimise. In addition to cleaner code this will enable further changes e.g. shared data reads for facet groups where reading needs to be called once for a group but processing needs to be called for each facet.

This change also updates some outdated documentation and removes an unused `rowId` member from the `InversionRow` struct