PAN-211 Move MS data initialisation
The overall goal of this commit is to reach a point where we are able to do shared "MSDataVector" initialisation of multiple gridders at once rather than having to repeat the initialisation for each gridder.
This commit enables this while also doing a lot of n ecessarry refactoring in order to not only make this change as clean as possible API wise but also to enable other similar changes that we are going to need to make for the full shared read functionality.

This is a stepping stone to full shared reads that gets us a lot of the way there but not the whole way, for the sake of keeping MRs manageable it makes sense to do some further changes in future MRs rather than all at once.

This also helps remove some redundancy from the codebase, e.g. instead of every class that derives from `MSGridderBase` calling `initializeMSDataVector` the equivalent of this now only gets called a single time by the gridder manager.

Main changes:
* Move `initializeMSDataVector` and associated data like `MSData` struct out of `MSGridderBase` and into `MSGridderManager`
* Also move all management of `MSProvider` out of `MSGridderBase` and into `MSGridderManager`
* Introduce new `MSManager` class
    - To keep the above changes clean, rather than adding it all directly to `MSGridderManager`  we create a new class `MSManager` that becomes responsible for handling all `MSData`/`MSProvider` related functionality in terms of managing and initialising this data.
    - `MSGridderManager` gets an `MSManager` member and calls the appropriate initialisation calls etc.
* Split `MSData` struct into two `Data` and `FacetData`
    - `Data` contains shared data for a MS that is the same across all facets in a facet group
    - `FacetData` contains facet specific data for a MS that is unique to each facet and can't be shared across other facets in the facet group
* `Data` and `FacetData` move out of `MSGridderBase` and become subtypes of `MSManager` which also contains instances to manage this data
* A third type `Limits` is also introduced which takes over some variables that were previously populated by `calculateMSLimits`
    - `Limits` is also shared across facets in a group but is different from `Data` because there is one `Data` per MS but a single `Limits` for multiple MS
    - `Limits` also becomes a subtype of `MSManager`
    - `calculateMSLimits` is replaced by a `Calcuate` method on the `Limits` struct allowing this code to be cleaner and more concise
* Take advantage of the new placing of data to use const in more places when passing them as arguments
* Temporarily introduce a new method `initializeVisibilityModifierTimes`
    - This method is necessarry to handle some visibility modifier related initialisation that used to be part of the now removed `MSGridderBase::initializeMSDataVector`
    - As this method also reads through the data for each gridder instead of doing a shared read it also needs to (in future) be refactored to read through the data only once, and likely some of this behaviour belongs in `MSGridderManager`
    - We leave this to be done in a subsequent MR as otherwise this MR starts to become too bloated and complex