Fix issues with running nightlychecks twice in a folder
Some of the tests that are run by `make nightlychecks` modify the `.ms` they run on. This creates a situation where if you run `make nightlychecks` multiple times in the same data directory the tests will fail on all runs other than the first one. This is fine where running the tests on a CI system but when trying to debug test failures or develop new tests it is a bit of a pain.

Fix this by changing the tests in question to work on a temporary copy instead of the source ms