PAN-211 Alter wslimit behaviour for facet groups
Review discussion in MR !717 introduced the idea that we can both simplify and speed up the code a bit by calculating wlimits once for the smallest facet in a facet group.
And then use those values for all facets instead of calculating individually for each facet.
This commit introduces that behaviour.

Some explanation extracted from the code:
```
  // wlimits will vary across facets in a facet group, however these limits are
  // "estimates".
  // They should not be too small but can be too large (though this can have
  // performance implications.
  // As a code simplification and performance improvement we select one facet in
  // a facet group to calculate the wlimits rather than doing it individually
  // for each one.
  // We select the smallest facet for this as width/height of a facet influence
  // the uv-grid size inversely and we want the largest uv-grid as that is
  // likely to have more extreme w values
```