PAN-211 Rename member variables of new class/struct
In MR !717 a lot of code was moved out of MSGridderBase and into new classes.
Some of the member variables that moved are from old code that did not follow the coding style guideline.
As it was quite a sizeable MR these variables were left as is instead of updated to the new style to avoid bloating the MR further.

This is a follow up MR to do those renames, this commit purely refactors variable names to program behaviour is altered.