PAN-233 Split parts of MSGridderBase into MsGridderData
Shared reads need to use various methods (ApplyCorrections, CalculateWeights etc.) that are currently in MSGridderBase without a gridder, because the manager will do these things on shared data outside of the gridder.
To allow for this it becomes necessary for a relatively large portion of things that are currently part of MSGridderBase to be moved out of it.

Composition was considered/attempted but was not working well; it made the rest of the code very clunky plus may have had some performance ramifications...
So we instead use a new base class `MsGridderData` from which `MSGridderBase` inherits and move a lot of code out of `MSGridderBase` into `MsGridderData`

This also has the benefit of slightly cleaning up the structure of the codebase as `MSGridderBase` is too large and doing too many things, so this makes some good steps in seperating this a bit.

This commit has no functional changes, changes are almost entirely limited to moving code between classes, though there are one or two very minor code alterations to better accommodate this.
Also some renaming to update to current coding styles is done.