PAN-233 Explicitely init cfitsio to remove race condition
If you don't explicitly init cfitsio then cfitsio does automatic initialisation inside the first call to `fits_open_file` (or one of several other possible calls).

However it uses a global boolean that is not mutex protected to do this, so if multiple threads (for example) call `fits_open_file` at the same time, then a race occurs that can lead to state corruption and a crash.

While its not clear if any of the current code in WSClean can actually trigger this condition I did at some point during development write some code that did trigger it and as the fix is simple it seems better to rather have it in just in case.